refactor: breakout FieldSelector into separate components
